Fix images in base64 on activity content blocks and refactor methods.
